+++ programs/drift/src/instructions/user.rs
@@ -154,6 +154,9 @@ pub fn handle_initialize_user_stats(ctx: Context<InitializeUserStats>) -> Result
     Ok(())
 }
 
+#[access_control(
+    deposit_not_paused(&ctx.accounts.state)
+)]
 pub fn handle_deposit(
     ctx: Context<Deposit>,
     market_index: u16,
@@ -472,6 +475,7 @@ pub fn handle_withdraw(
 }
 
 #[access_control(
+    deposit_not_paused(&ctx.accounts.state)
     withdraw_not_paused(&ctx.accounts.state)
 )]
 pub fn handle_transfer_deposit(